create ingester classes
